# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow or unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
  /users/{userId}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
  /users/me/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Chat
...
  /services/assignments:
    get:
      operationId: getAssignments
      tags: [services]
      summary: Get assignments list
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Assignments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentsResponse"

    post:
      operationId: createAssignment
      tags: [services]
      summary: Create assignment request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /services/assignments/{assignmentId}/accept:
    post:
      operationId: acceptAssignment
      tags: [services]
      summary: Accept an assignment
      security:
        - bearerAuth: []
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assignment accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  # Services - Coaching
  /services/coaching:
    get:
      operationId: getCoachingSessions
      tags: [services]
      summary: Get coaching sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSessionsResponse"

    post:
      operationId: createCoachingSession
      tags: [services]
      summary: Create coaching session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoachingSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSession"
  /services/coaching/{sessionId}/book:
    post:
      operationId: bookCoachingSession
      tags: [services]
      summary: Book a coaching session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSession"

  # Services - Delivery
...
  /services/tasks:
    get:
      operationId: getTasks
      tags: [services]
      summary: Get task listings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResponse"

    post:
      operationId: createTask
      tags: [services]
      summary: Post a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /services/tasks/{taskId}/apply:
    post:
      operationId: applyForTask
      tags: [services]
      summary: Apply for a task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        college:
          type: string
          nullable: true
        program:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, provider, admin]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - followersCount
        - followingCount
        - postsCount
        - createdAt
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            isFollowing:
              type: boolean
          required:
            - isFollowing
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        college:
          type: string
        program:
          type: string
        avatar:
          type: string
    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required:
        - following
        - followersCount

    # Post schemas
    Post:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        author:
          $ref: "#/components/schemas/User"
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - mediaUrls
        - author
        - likesCount
        - commentsCount
        - isLiked
        - createdAt
    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: string
          nullable: true
      required:
        - posts
...
    Assignment:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryMode:
          type: string
          enum: [pdf, zip, physical]
        status:
          type: string
          enum: [open, accepted, completed, cancelled]
        poster:
          $ref: "#/components/schemas/User"
        acceptedBy:
          $ref: "#/components/schemas/User"
          nullable: true
        subject:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - price
        - deliveryMode
        - status
        - poster
        - createdAt
    AssignmentsResponse:
      type: object
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"
      required:
        - assignments
    CreateAssignmentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryMode:
          type: string
          enum: [pdf, zip, physical]
        subject:
          type: string
        deadline:
          type: string
          format: date-time
      required:
        - title
        - description
        - price
        - deliveryMode
    CoachingSession:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        sessionType:
          type: string
          enum: [one_on_one, group]
        subject:
          type: string
        mentor:
          $ref: "#/components/schemas/User"
        bookedBy:
          $ref: "#/components/schemas/User"
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [available, booked, completed]
        maxStudents:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - price
        - sessionType
        - subject
        - mentor
        - status
        - maxStudents
        - createdAt
    CoachingSessionsResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/CoachingSession"
      required:
        - sessions
    CreateCoachingSessionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        sessionType:
          type: string
          enum: [one_on_one, group]
        subject:
          type: string
        scheduledAt:
          type: string
          format: date-time
        maxStudents:
          type: integer
      required:
        - title
        - description
        - price
        - sessionType
        - subject
        - maxStudents
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        category:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        poster:
          $ref: "#/components/schemas/User"
        assignedTo:
          $ref: "#/components/schemas/User"
          nullable: true
        applicantsCount:
          type: integer
        deadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - budget
        - category
        - status
        - poster
        - applicantsCount
        - createdAt
    TasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required:
        - tasks
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        category:
          type: string
        deadline:
          type: string
          format: date-time
      required:
        - title
        - description
        - budget
        - category

    # Notification schemas
...