# lib/api-zod/src/generated/types/coachingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachingSessionSessionType } from "./coachingSessionSessionType";
import type { CoachingSessionStatus } from "./coachingSessionStatus";
import type { User } from "./user";

export interface CoachingSession {
  id: string;
  title: string;
  description: string;
  price: number;
  sessionType: CoachingSessionSessionType;
  subject: string;
  mentor: User;
  bookedBy?: User | null;
  scheduledAt?: Date | null;
  status: CoachingSessionStatus;
  maxStudents: number;
  createdAt: Date;
}